Add Footer, NewGraphElement, CarouselElement components and diagramType enum
