+++ diagramflow/frontend/src/components/Footer.tsx
@@ -1,0 +1,14 @@
+export function Footer() {
+  return (
+    <div className="bg-gray-200 text-center text-sm text-gray-600 py-2 border-t rounded-2xl flex flex-row justify-between">
+      <div className="mx-8">
+        <span className="font-bold text-blue-500">flow</span>
+        <span className="text-black">draw.</span>
+      </div>
+      <div>
+        Copyright © 2025 FlowDraw. Wszelkie prawa zastrzeżone.
+      </div>
+      <div className="mx-16"/>
+    </div>
+  )
+}
