+++ diagramflow/frontend/src/types/diagramType.ts
@@ -1,0 +1,5 @@
+export enum diagramType {
+  UML,
+  Flowchart,
+  Network,
+}
